+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# emphatic 0.1.6.9010  2024-05-02
+# emphatic 0.1.6.9011  2024-05-04
 
 * Removed `hl_mat()` and associated vignettes.
     * function is not exported for now as the selection process is too 
@@ -14,6 +14,8 @@
     * Output to raw blocks in typst
     * only character to be escaped is the backtick
 * latex rendering now has better whitespace at start of line
+* SVG anim support
+* Excel export
 
 # emphatic 0.1.6  2024-04-27